Component division, ready to update rest requesting
