+++ front-end/src/Views/RegisterPage/functions/handlers.js
@@ -1,5 +1,7 @@
 import {Redirect} from "react-router-dom";
 import React from "react";
+import HomeDataForm from "../components/homeDataForm";
+import CompanyDataForm from "../components/companyDataForm";
 
 export const onChange = (component, e, val) => {
     const value = e.target.value;
@@ -20,3 +22,11 @@ export const setRedirect = component => {
         redirect: true
     });
 };
+
+export const renderSection = (component) => {
+    switch(component.state.account_type) {
+        case "Podopiecznym": {return <HomeDataForm component={component} />}; break;
+        case "Pracodawcą": {return <CompanyDataForm component={component} /> }; break;
+        case "Administratorem": {return null;}; break;
+    }
+}
